refactor : define a generic method to upload image to storage
