+++ app/src/main/java/com/android/solvit/shared/model/utils/ImageLoader.kt
@@ -4,6 +4,9 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.ImageDecoder
 import android.net.Uri
+import com.google.firebase.storage.FirebaseStorage
+import com.google.firebase.storage.StorageReference
+import java.util.UUID
 
 // Helper function to load a Bitmap from a URI
 fun loadBitmapFromUri(context: Context, uri: Uri): Bitmap? {
@@ -14,4 +17,24 @@ fun loadBitmapFromUri(context: Context, uri: Uri): Bitmap? {
     e.printStackTrace()
     null
   }
+}
+
+fun uploadImageToStorage(
+    storage: FirebaseStorage,
+    path: String,
+    imageUri: Uri,
+    onSuccess: (String) -> Unit,
+    onFailure: (Exception) -> Unit
+) {
+  val uniqueFileName = UUID.randomUUID().toString() + ".jpg"
+  val imageRef: StorageReference = storage.reference.child(path + uniqueFileName)
+
+  imageRef
+      .putFile(imageUri)
+      .addOnSuccessListener {
+        imageRef.downloadUrl
+            .addOnSuccessListener { downloadUrl -> onSuccess(downloadUrl.toString()) }
+            .addOnFailureListener { exception -> onFailure(exception) }
+      }
+      .addOnFailureListener { exception -> onFailure(exception) }
 }
